refactor, removed catch-all py exceptions (doesn't work) moved statics.js into structs.js which later will be refactored again
