+++ .gitignore
@@ -14,3 +14,5 @@ update.log
 time.log
 debug.log
 performance_time.log
+error.log
+exception-logger.log
