+++ modules/python/utils.py
@@ -31,17 +31,7 @@ err_formatter = logging.Formatter(logging.BASIC_FORMAT)
 err_handler.setFormatter(err_formatter)
 
 err_logger = logging.getLogger("error-logger")
-err_logger.setLevel(logging.WARNING)
 err_logger.addHandler(err_handler)
-
-default_uncaught = sys.excepthook
-
-# Add logging to debug.log for uncaught exceptions
-def exception_handler(type, value, traceback):
-    err_logger.error("[Uncaught exception] {}: {}\nTraceback:\n{}".format(type, value, traceback))
-    default_uncaught(type, value, traceback)
-
-sys.excepthook = exception_handler
 
 def getFunctionBlock(frame) -> gdb.Block:
     block = frame.block()
